refactor(views): rename TimelineView → ScheduleView; delete dead ScheduleView
The toolbar shows "Schedule" and the view id is 'schedule', but the
component rendered was named TimelineView. That triple-name mismatch
made it surprisingly easy to read code wrong (see the walkthrough
debugging session: "TimelineView's rows are employees" was correct but
required cross-referencing two names to grasp).

Changes:

  • git mv src/views/TimelineView.tsx       → src/views/ScheduleView.tsx
  • git mv src/views/TimelineView.module.css → src/views/ScheduleView.module.css
  • git mv 4 TimelineView.*.test.tsx       → ScheduleView.*.test.tsx
    (touchDnd, dayWindow, baseFilterRecovery, grouping)
  • Component / interface / variable renamed inside the file
  • CSS import path updated
  • src/index.ts re-export renamed (`TimelineView` → `ScheduleView`)
  • src/WorksCalendar.tsx import + JSX usage renamed
  • src/ui/__tests__/a11y.test.tsx import + describe + variable renamed
  • src/__tests__/WorksCalendar.employees.sync.test.tsx test name
    updated
  • Comment / docstring references updated in: ApprovalDot, EmployeeActionCard,
    DayWindowPills, useCalendar, AssetsView (.tsx + .module.css),
    BaseGanttView (.tsx + .module.css), examples/10-DragAndDrop.tsx
  • New ScheduleView.tsx top-of-file comment notes the historical
    rename so anyone grepping "TimelineView" hits a breadcrumb

Dead-code removal (no production references — only its own test file):
  • Delete src/views/ScheduleView.tsx (the unused 6-week grid)
  • Delete src/views/ScheduleView.module.css
  • Delete src/views/__tests__/ScheduleView.test.tsx (13 tests for
    the deleted component)

The view id stays 'schedule' so user-facing label and config defaults
(display.defaultView) are unaffected.

Verified:
  • tsc --noEmit clean
  • vitest: 186 test files / 2649 tests pass (was 187 / 2662 — the
    delta of 1 file / 13 tests is exactly the deleted dead-component
    suite, no live regressions)
  • demo build succeeds
  • zero remaining 'TimelineView' references in src/, demo/, examples/
    except one deliberate breadcrumb comment in the renamed file
